pyproject.toml
[project]
name = "claude-docker-sandbox"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
